requirements.txt
aiosmtplib # async emails

openai # OpenAI SDK

python-dotenv # env

psycopg2 # postgres

openpyxl # xlsx

httpx # proxy-capable HTTP client

pyodbc # SQL Server via ODBC
